services/rag_services/requirements.txt
# Core dependencies for Vietnamese Hybrid RAG system
fastapi==0.117.1
uvicorn==0.37.0
pydantic==2.11.9
pydantic-settings==2.11.0

# OpenSearch and BM25 dependencies
opensearch-py==2.4.2
rank-bm25==0.2.2

# Weaviate vector database - self-hosted
weaviate-client==4.9.3

# ML and embeddings
sentence-transformers==5.1.1
transformers==4.56.2
torch==2.8.0

# Data processing
numpy>=1.24.0
pandas==2.2.3
scikit-learn==1.7.2

# PDF processing
PyPDF2==3.0.1

# HTTP clients and utilities
requests==2.32.5
httpx==0.27.0
aiohttp==3.12.15

# Additional utilities
python-dotenv==1.1.1
PyYAML==6.0.2
tqdm==4.67.1

# Week 2 dependencies
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0  # For faster fuzzy matching

# LLM providers for relation extraction
openai>=1.0.0  # OpenAI GPT-4
google-generativeai>=0.3.0  # Google Gemini

